Cosmetics and coding style fixes in devices/*.
 - Whitespace and indentation fixes in various places.
 
 - Fix various typos.
 
 - Use u8, u16 etc. everywhere.

Abuild-tested.

Acked-by: Uwe Hermann <uwe@hermann-uwe.de>



git-svn-id: svn://svn.coreboot.org/coreboot/trunk@5962 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1

+++ src/include/device/pnp.h
@@ -5,15 +5,15 @@
 #include <device/device.h>
 #include <device/pnp_def.h>
 
-/* Primitive pnp resource manipulation */
-void    pnp_write_config(device_t dev, uint8_t reg, uint8_t value);
-uint8_t pnp_read_config(device_t dev, uint8_t reg);
-void    pnp_set_logical_device(device_t dev);
-void    pnp_set_enable(device_t dev, int enable);
-int     pnp_read_enable(device_t dev);
-void    pnp_set_iobase(device_t dev, unsigned index, unsigned iobase);
-void    pnp_set_irq(device_t dev, unsigned index, unsigned irq);
-void    pnp_set_drq(device_t dev, unsigned index, unsigned drq);
+/* Primitive PNP resource manipulation */
+void pnp_write_config(device_t dev, u8 reg, u8 value);
+u8 pnp_read_config(device_t dev, u8 reg);
+void pnp_set_logical_device(device_t dev);
+void pnp_set_enable(device_t dev, int enable);
+int pnp_read_enable(device_t dev);
+void pnp_set_iobase(device_t dev, u8 index, u16 iobase);
+void pnp_set_irq(device_t dev, u8 index, u8 irq);
+void pnp_set_drq(device_t dev, u8 index, u8 drq);
 
 /* PNP device operations */
 void pnp_read_resources(device_t dev);
@@ -31,8 +31,8 @@ struct io_info {
 
 struct pnp_info {
 	struct device_operations *ops;
-	unsigned function;
-	unsigned flags;
+	unsigned int function;
+	unsigned int flags;
 #define PNP_IO0  0x001
 #define PNP_IO1  0x002
 #define PNP_IO2  0x004
@@ -48,6 +48,6 @@ struct pnp_info {
 };
 struct resource *pnp_get_resource(device_t dev, unsigned index);
 void pnp_enable_devices(struct device *dev, struct device_operations *ops,
-	unsigned functions, struct pnp_info *info);
+			unsigned int functions, struct pnp_info *info);
 
 #endif /* DEVICE_PNP_H */
